+++ genium/src/test/resources/smoke/basic/output/cpp/include/smoke_off/Basic.h
@@ -11,7 +11,7 @@
 
 #include <string>
 
-namespace smoke {
+namespace smoke_off {
 
 class Basic {
 public: